pyproject.toml
[tool.pytest.ini_options]
pythonpath = [
    ".", "src",
]
asyncio_mode="auto"




[tool.poetry]
name = "microchat"
version = "0.1.0"
description = "Small messanger"
authors = ["SosisterRapStar <timoshenkovvanya@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standart"], version = "^0.27.0"}
fastapi = "^0.109.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
pydantic-settings = "^2.1.0"
asyncpg = {extras = ["standart"], version = "^0.29.0"}
alembic = {extras = ["standart"], version = "^1.13.1"}
websockets = {extras = ["standart"], version = "^12.0"}
pydantic = {extras = ["email"], version = "^2.6.0"}
python-multipart = "^0.0.9"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.2"
aioredis = "^2.0.1"
requests = "^2.31.0"
httpx = "^0.27.0"
anyio = "^4.3.0"
pytest-asyncio = "^0.23.6"
docker = "^7.0.0"
python-dotenv = "0.21.1"
faker = "^24.4.0"
nest-asyncio = "^1.6.0"
polyfactory = "^2.15.0"
pytest-env = "^1.1.3"
pytest-dotenv = "^0.5.2"
celery = "^5.3.6"
flower = "^2.0.1"
redis = "^5.0.3"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
    